Move type metadata and exhaustive primary key tests to test module
